cart

+++ src/main/java/com/teama/irenew/repository/ProductRepo.java
@@ -3,7 +3,7 @@ package com.teama.irenew.repository;
 import com.teama.irenew.models.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepo extends JpaRepository<Product, Long> {
+public interface ProductRepo extends JpaRepository<Product, Integer> {
 }
